requirements.txt
Jinja2==3.1.2
multiping==1.1.2
influxdb_client==1.36.1
python-dotenv==1.0.0